+++ test/browser/control.calendar.test.js
@@ -22,18 +22,22 @@ describe("control.calendar", () => {
         const el = document.querySelector("wup-calendar");
         return { headerText: el.querySelector("header>button").textContent, h: el.offsetHeight, w: el.offsetWidth };
       });
-    let info = await getInfo();
     const sz = { h: 262, w: 246 };
-    expect(info).toEqual({ headerText: "September 2022", ...sz });
+    expect(await getInfo()).toEqual({ headerText: "September 2022", ...sz });
+    // scroll to check how it can affect on size
+    await page.evaluate(() => {
+      const ol = document.querySelector("wup-calendar ol");
+      ol.dispatchEvent(new WheelEvent("wheel", { bubbles: true, cancelable: true, deltaY: 100 })); // scrollDown
+    });
+    await page.waitForTimeout(500);
+    expect(await getInfo()).toEqual({ headerText: "October 2022", ...sz });
 
     await page.click("wup-calendar header>button");
     await page.waitForTimeout(310);
-    info = await getInfo();
-    expect(info).toEqual({ headerText: "2022", ...sz });
+    expect(await getInfo()).toEqual({ headerText: "2022", ...sz });
 
     await page.click("wup-calendar header>button");
     await page.waitForTimeout(310);
-    info = await getInfo();
-    expect(info).toEqual({ headerText: "2018 ... 2033", ...sz });
+    expect(await getInfo()).toEqual({ headerText: "2018 ... 2033", ...sz });
   });
 });